+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ----------------------------------------
 
+## [1.7.0] - 2026-01-17 [DEVICE-COLOR]
+
+### Added
+- Add per-device color setting to tint device tabs and show a swatch in device lists.
+
+### Changed
+- None.
+
+### Deprecated
+- None.
+
+### Removed
+- None.
+
+### Fixed
+- None.
+
+### Security
+- None.
+
+----------------------------------------
+
 ## [1.6.1] - 2026-01-17 [SFTP-QUICK-SEARCH]
 
 ### Added